Bug 1980348 - fix test that was giving non-deterministic results.
@Ignore-d this in the previous patch, so i'm fixing that separately.
This test manipulates Tab objects, which are being transitioned between active and inactive states by a time-dependent Flow.
We're testing the storage of these objects, so i've pulled out the logic to compare tabs to its own assertTabSame() function, which looks at the contents of the tabs rather than whether they're currently active.

Also don't think we need to use spy() here for the class under test, so i've dropped in the real thing.

Differential Revision: https://phabricator.services.mozilla.com/D264246